requirements.txt
aerich==0.9.1
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
dictdiffer==0.9.0
dotenv==0.9.9
fastapi==0.116.1
filelock==3.18.0
firebase_admin==7.1.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-documentai==3.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-genai==1.29.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pypika-tortoise==0.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tortoise-orm==0.25.1
tqdm==4.67.1
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
